+++ API.Core/Entities/Literature.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace API.Entities
+namespace API.Core.Entities
 {
     public class Literature
     {
