+++ API.Core/Entities/Subject.cs
@@ -3,14 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace API.Entities
+namespace API.Core.Entities
 {
-    public class Author
+    public class Subject
     {
         public int Id { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public DateTime DateOfBirth { get; set; }
+        public string Name { get; set; }
 
         public ICollection<Literature> Literatures { get; set; }
     }
