+++ API.Data/Data/MapperProfileAPI.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using AutoMapper;
+
+namespace API.Data.Data
+{
+    public class MapperProfileAPI : Profile
+    {
+        public MapperProfileAPI()
+        {
+            
+        }
+    }
+}
