+++ API/Startup.cs
@@ -11,6 +11,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using API.Data.Data;
+using Microsoft.EntityFrameworkCore;
 
 namespace API
 {
@@ -32,6 +34,10 @@ namespace API
             {
                 c.SwaggerDoc("v1", new OpenApiInfo { Title = "API", Version = "v1" });
             });
+
+            services.AddDbContext<DbContextAPI>(options => options.UseSqlServer(Configuration.GetConnectionString("DbContextAPI")));
+
+            services.AddAutoMapper(typeof(MapperProfileAPI));
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
