+++ API/appsettings.json
@@ -6,5 +6,8 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "DbContextAPI": "Server=(localdb)\\mssqllocaldb;Database=DbContextAPI;Trusted_Connection=True;MultipleActiveResultSets=true"
+  }
 }
